Changes to be committed     modified:   index.html     modified:   prompt.json
I added a new category to the prompts.
I also added a new prompt to the every day category.
I also added a new prompt to the deep category.
I also added a new prompt to the creative category.
I also added a new prompt to the funny category.
I also added a new prompt to the thought-provoking category.